diabetes_xai/requirements.txt
streamlit>=1.50.0
pandas>=2.0.0
numpy>=1.26.0,<3.0.0
scikit-learn>=1.4.0
shap>=0.45.0
plotly>=5.20.0
matplotlib>=3.8.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
fairlearn>=0.10.0
joblib>=1.3.0
